refactor: 사용자 입력 도메인, 유틸로 분리
getCarsName, getGameCount는 도메인이므로 input.js, readLineAsync는 utils.js로 분리
