+++ src/View/utils.js
@@ -1,0 +1,25 @@
+import readline from 'readline';
+
+function readLineAsync(query) {
+  return new Promise((resolve, reject) => {
+    if (arguments.length !== 1) {
+      reject(new Error('arguments must be 1'));
+    }
+
+    if (typeof query !== 'string') {
+      reject(new Error('query must be string'));
+    }
+
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(query, (input) => {
+      rl.close();
+      resolve(input);
+    });
+  });
+}
+
+export default readLineAsync;
